+++ packages/ui/cypress/integration/Regression/CreateModels/EntityLabeling.spec.js
@@ -56,11 +56,12 @@ describe('Entity Labeling - Create Model', () => {
 
     it('Label multiple words as the same entity.', () => {
       train.TypeYourMessage('This is Frog and Tag.')
-      memoryTableComponent.VerifyEntityValues('multi', ['Tag'])
+      memoryTableComponent.VerifyExistingEntityValues('multi', ['Tag'])
       entityDetectionPanel.LabelTextAsEntity('Tag', 'multi', 0, false)
       entityDetectionPanel.LabelTextAsEntity('Frog', 'multi', 0, false)
       train.ClickScoreActionsButton()
-      memoryTableComponent.VerifyEntityValues('multi', ['Tag', 'Frog'])
+      memoryTableComponent.VerifyExistingEntityValues('multi', ['Tag'])
+      memoryTableComponent.VerifyNewEntityValues('multi', ['Frog'])
       train.SelectTextAction('Hi')
     })
 
@@ -73,7 +74,7 @@ describe('Entity Labeling - Create Model', () => {
     // Bug 2351: Inconsistant Automatic Entity Labeling
     it('Reverse the labeled words and once again label them as the same entity.', () => {
       train.TypeYourMessage('This is Tag and Frog.', [{ text: 'Tag', entity: 'multi' }, { text: 'Frog', entity: 'multi' }])
-      memoryTableComponent.VerifyEntityValues('multi', ['Tag', 'Frog'])
+      memoryTableComponent.VerifyExistingEntityValues('multi', ['Tag', 'Frog'])
       entityDetectionPanel.VerifyTextIsLabeledAsEntity('Tag', 'multi')
 
       // Sometimes it is labeled, sometimes it is not - bug 2351
@@ -81,6 +82,7 @@ describe('Entity Labeling - Create Model', () => {
       //entityDetectionPanel.VerifyTextIsLabeledAsEntity('Frog', 'multi') // this is the line of code we should need
 
       train.ClickScoreActionsButton()
+      memoryTableComponent.VerifyExistingEntityValues('multi', ['Tag', 'Frog'])
       train.SelectTextAction('Hi')
     })