+++ packages/ui/cypress/support/MonitorDocumentChanges.js
@@ -4,12 +4,12 @@
 * 
 * AUTHOR: Michael Skowronski
 *
-* This code eliminiates the need for adding cy.wait() commands to your code and also
+* This code eliminiates most of the need for adding cy.wait() commands to your code and also
 * cy.route() commands used only for waiting. It does this by constantly monitoring the
-* DOM for changes and tracking the Milliseconds Since the Last Change and also resetting
-* those Milliseconds to zero when we see the spinner is displayed. Certain cy.commands(),
-* such as cy.Get(), are then modified to prevent execution until this Millisecond count
-* reaches at least 700.
+* DOM for the spinner and for changes. It tracks the Milliseconds Since the Last Change and 
+* resets those Milliseconds to zero when we see the spinner is displayed. Certain cy.commands(),
+* such as cy.get(), are then modified and renamed (cy.Get()) to prevent execution until this 
+* Millisecond count reaches at least 700.
 * 
 * The basic premis this works on is that when the application under test is making API
 * calls, the spinner is displayed, therefore further test steps should be paused. Also when 
@@ -23,7 +23,6 @@ import * as helpers from './Helpers.js'
 (function () {
   let lastMonitorTime = 0
   let lastChangeTime = 0
-  let waitTillNChangesOccur = 0
   let lastHtml
   let StopLookingForChanges = false
   let dumpHtml = false
@@ -35,16 +34,6 @@ import * as helpers from './Helpers.js'
 
   function MillisecondsSinceLastChange() {
     if (new Date().getTime() > lastMonitorTime + 50) LookForChange()
-
-    // Some operations are known to take a long time without rendering, so while this count
-    // is greater than zero, we will reset our lastChangeTime and return zero which will
-    // effectively cause a wait until the specified number of changes pass and then resume
-    // our normal monitoring for stable DOM.
-    if (waitTillNChangesOccur > 0) {
-      lastChangeTime = new Date().getTime()
-      return 0
-    }
-
     return (new Date().getTime() - lastChangeTime)
   }
 
@@ -147,27 +136,6 @@ import * as helpers from './Helpers.js'
         })
     })
 
-    // This will force any of our code that waits for stable DOM to wait until 'changeCount' number of changes
-    // are detected, and then another 700 milliseconds of no changes (resuming normal operation) before the 
-    // DOM is considered stable.
-    //
-    // The use case for this is when the normal waiting fails because we are expecting a change to the DOM but 
-    // it takes too long to occur. This is typically a product bug, but if not, or to temporarily workaround the
-    // product bug you can consult the logs and see what changes come and determine the count of those changes
-    // that must occur before normal operations are to resume.
-    Cypress.Commands.add('WaitTillNChangesOccur', { prevSubject: 'optional' }, (elements, changeCount) => {
-      helpers.ConLog(`cy.WaitTillNChangesOccur(${changeCount})`, `Start - Last DOM change was ${MillisecondsSinceLastChange()} milliseconds ago`)
-      waitTillNChangesOccur = changeCount
-    })
-
-    Cypress.Commands.add('RunAndExpectDomChange', (func) => {
-      helpers.ConLog(`cy.RunAndExpectDomChange()`, `Start - Last DOM change was ${MillisecondsSinceLastChange()} milliseconds ago`)
-      lastChangeTime = new Date().getTime()
-      func()
-      helpers.ConLog(`cy.RunAndExpectDomChange()`, `done - Last DOM change was ${MillisecondsSinceLastChange()} milliseconds ago`)
-      lastChangeTime = new Date().getTime()
-    })
-
     Cypress.Commands.add('DumpHtmlOnDomChange', (boolValue) => { dumpHtml = boolValue })
 
     Cypress.on('window:before:load', () => {
@@ -190,9 +158,9 @@ import * as helpers from './Helpers.js'
     setTimeout(() => { LookForChange(true) }, 50)   // This will repeat until Stop is called.
 
     helpers.ConLog(`MonitorDocumentChanges.initialize()`, `Running`)
-  }
+  } // end of function Initialize()
 
-  let dumpSpinner = false
+
   function LookForChange(loop) {
     const funcName = `MonitorDocumentChanges.LookForChange(${loop})`
 
@@ -204,13 +172,12 @@ import * as helpers from './Helpers.js'
     const currentTime = lastMonitorTime = new Date().getTime()
     const currentHtml = Cypress.$('html')[0].outerHTML
     if (currentHtml != lastHtml) {
-      if (waitTillNChangesOccur > 0) { waitTillNChangesOccur-- }
       helpers.ConLog(funcName, `Change Found - Milliseconds since last change: ${(currentTime - lastChangeTime)}`)
       if (dumpHtml) {
         helpers.ConLog(funcName, `Current HTML:\n${currentHtml}`)
       } else {
-        // If the error panel is up dump its elements to the log file so that they can be used to help
-        // identify a bug using our test result evaluation tool.
+        // If the error panel is showing, dump its elements to the log file so that they can be used to help
+        // identify a bug using our Test Result Evaluation Tool.
         const errorPanelElements = Cypress.$('div.cl-errorpanel')
         if (errorPanelElements.length > 0) {
           helpers.DumpElements(`${funcName} - Error Panel found in Current HTML`, errorPanelElements)