+++ packages/models/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.502.0](https://github.com/Microsoft/conversationlearner-models/compare/v0.501.17...v0.502.0) (2020-01-13)
+
+
+### Features
+
+* action callback results - phase 1 ([#27](https://github.com/Microsoft/conversationlearner-models/issues/27)) ([552f15b](https://github.com/Microsoft/conversationlearner-models/commit/552f15b971caac78f0b9adf4e712874826e2e9d2))
+
+
+
+
+
 ## [0.501.2](https://github.com/Microsoft/conversationlearner-models/compare/v0.501.1...v0.501.2) (2019-12-19)
 
 **Note:** Version bump only for package @conversationlearner/models