+++ packages/samples/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.502.1](https://github.com/microsoft/conversationlearner/compare/v0.502.0...v0.502.1) (2020-01-18)
+
+**Note:** Version bump only for package conversationlearner-samples
+
+
+
+
+
+# [0.502.0](https://github.com/microsoft/conversationlearner/compare/v0.501.17...v0.502.0) (2020-01-13)
+
+
+### Features
+
+* action callback results - phase 1 ([#27](https://github.com/microsoft/conversationlearner/issues/27)) ([552f15b](https://github.com/microsoft/conversationlearner/commit/552f15b971caac78f0b9adf4e712874826e2e9d2))
+
+
+
+
+
+## [0.501.17](https://github.com/microsoft/conversationlearner/compare/v0.501.16...v0.501.17) (2020-01-10)
+
+**Note:** Version bump only for package conversationlearner-samples
+
+
+
+
+
+## [0.501.16](https://github.com/microsoft/conversationlearner/compare/v0.501.15...v0.501.16) (2020-01-09)
+
+**Note:** Version bump only for package conversationlearner-samples
+
+
+
+
+
 ## [0.501.15](https://github.com/microsoft/conversationlearner/compare/v0.501.14...v0.501.15) (2020-01-03)
 
 **Note:** Version bump only for package conversationlearner-samples