+++ packages/webchat/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.502.0](https://github.com/Microsoft/ConversationLearner-WebChat/compare/v0.501.17...v0.502.0) (2020-01-13)
+
+**Note:** Version bump only for package @conversationlearner/webchat
+
+
+
+
+
 # [0.501.0](https://github.com/Microsoft/ConversationLearner-WebChat/compare/v0.500.3...v0.501.0) (2019-12-18)
 
 **Note:** Version bump only for package @conversationlearner/webchat